Features : strong password,encrypted user token using digital signature
added validation at front end and backend for password. password should be greater than 8 characters and it should contain atleast one special character,one uppercase,one lowercase and one digit. Another added feature is that on login a encrypted user token is issued and returned to the user that can be used to verify user.Now user will not logout on reloading a page.